Minor change to improve support for Parcel/SWC
Parcel is currently [unable to perform static analysis](https://github.com/parcel-bundler/parcel/issues/7904) on the `in` operator applied to `process.env`, for the purposes of bundling the module. (Parcel is a front-end to SWC, so this most likely affects SWC as well.)

Would you mind making this very minor change?

It should work exactly the same - `enabled` is `true` by default, so `FORCE_COLOR` only has an effect when set to exactly `0`.

It's a bit simpler and avoids mutations, so maybe a bit easier to read anyhow. 🙂